rename OnyxProgressStep to OnyxProgressItem
